Use `request-promise` to avoid method name collisions
The new `gitlab` package uses `request-promise` which causes method name
collisions when in the same project as `request-promise-native`. This is
due to both of these packages attempting to attach the various Promise
methods such as `then`, `catch`, etc.

Once one of the `request-promise` implementations has configured `request`
it will proceed to error if another implementation tries to attach the same
methods again. To avoid any trouble, the default `request-promise`
implementation using Bluebird will be used (as it is not possible to set
the Promise implementation to be used by the `gitlab` API).