Add `User` model to encapsulate user identities
This commit adds a `User` model to properly encapsulate what data should
be exposed from the Staticman API and specifically narrow its scope.

With the addition of the GitLab identity provider, we need to be conscious
that the GitLab user data structure will not be the same as the GitHub
user (or any future identity provider's user model). This could be
problematic as:

1. We could expose private user data unintentionally.
2. Staticman consumers will be able too much access to the internal user
   API meaning they could become too reliant on parts that we do not want
   to expose long term.